MPC , tunning , contain speed , cost print info.

+++ Term2_Robotics_Sensor_Fusion/05_Project_MPC_Control/src/MPC.h
@@ -7,7 +7,7 @@
 using namespace std;
 
 extern double Lf;
-extern int dt;
+extern double dt;
 
 // This value assumes the model presented in the classroom is used.
 //
